requirements.txt
litserve>=0.1.0
openai-whisper
numpy
ffmpeg-python 
python-multipart